rename the SDK package to addon-sdk
